# api/doc/openapi-spec.yml
openapi: "3.0.3"
info:
  title: Meow Mingle social media app
  version: 0.0.1
servers:
  - url: http://127.0.0.1:8080/api/v1
...
paths:
...
  /posts/{id}:
    put:
      tags:
        - Posts
      description: Update post content
      operationId: updatePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            nullable: false
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostRequest"
      responses:
        200:
          description: Post updated successfully
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
    delete:
      tags:
        - Posts
      description: Delete post
      operationId: deletePost
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            nullable: false
      responses:
        204:
          description: Post deleted successfully
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
...
  /comments/{id}:
    put:
      tags:
        - Comments
      description: Update comment
      operationId: updateComment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentRequest"
      responses:
        200:
          description: Comment updated successfully
        400:
          $ref: "#/components/responses/400"
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
    delete:
      tags:
        - Comments
      description: Delete existing comment
      operationId: deleteCommentById
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
      responses:
        204:
          description: Comment deleted successfully
        401:
          $ref: "#/components/responses/401"
        403:
          $ref: "#/components/responses/403"
        404:
          $ref: "#/components/responses/404"
        500:
          $ref: "#/components/responses/500"
...
components:
  responses:
...
    400:
      description: Request not valid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    500:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: integer
    CommentRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 3
          maxLength: 300
          nullable: false
...
    PostRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 3
          maxLength: 300
          nullable: false
...